tests/test_base_agency.py
import unittest
from tasks import BaseTask
from rexia_ai.agents import ManagerAgent
from agencies import BaseAgency

class TestBaseAgency(unittest.TestCase):
    def setUp(self):
        self.manager = ManagerAgent()
        self.agents = []
        self.tasks = [BaseTask()]
        self.tools = []
        self.llm = None
        self.verbose = True
        self.agency = BaseAgency(self.manager, self.agents, self.tasks, self.tools, self.llm, self.verbose)

    def test_init(self):
        self.assertEqual(self.agency.manager, self.manager)
        self.assertEqual(self.agency.agents, self.agents)
        self.assertEqual(self.agency.tasks, self.tasks)
        self.assertEqual(self.agency.tools, self.tools)
        self.assertEqual(self.agency.llm, self.llm)
        self.assertEqual(self.agency.verbose, self.verbose)

    def test_launch(self):
        # Add test logic here
        pass

if __name__ == '__main__':
    unittest.main()